updated troy test cases
skipped the (showing validation when clicking next without answering) test that trung/troy had initially as it aligns with the codebase more, in questionnaire-form.tsx, line 558 it already renders it disabled as an attribute for if the question is answered initially. (if question is not answered, iscurrentquestionanswered() is returning false).

troy/trung test suite tests for TC37 and TC38 (completing the story config form and entering a scenario in the config page).

tldr: tc37 test verifies that form renders correctly with the first question visible, tests if submission is successful in form, which confirms the button labelled submit works. also checks that the toast message appears which verifies the confirmation message

tc38 test verifies the test cofnirms the questionnaire form page exists and renders, tests that the form can be submitted, and verifies form submission works correctly which includes the relevant api call to send to the admin panel.

mainly tests for rendering and submission like stated in discord. (core functionality)